# Improve labels computation during LogQL pipeline