This does two things:

- Introduce a buffer for computing labels and the buffer is only copied if we never seen that label set before.
- Add a Map function to the label builders used for `label_format` and `line_format` that avoids slice allocations and sorting.

```
❯ benchstat before.txt after.txt
name                                 old time/op    new time/op    delta
_Pipeline/pipeline_bytes-16            10.6µs ± 0%     6.9µs ± 1%  -34.85%  (p=0.008 n=5+5)
_Pipeline/pipeline_string-16           10.6µs ± 0%     6.9µs ± 0%  -35.16%  (p=0.008 n=5+5)
_Pipeline/line_extractor_bytes-16      10.6µs ± 0%     6.9µs ± 1%  -34.98%  (p=0.008 n=5+5)
_Pipeline/line_extractor_string-16     10.7µs ± 1%     6.9µs ± 0%  -35.27%  (p=0.016 n=5+4)
_Pipeline/label_extractor_bytes-16     10.7µs ± 1%     7.1µs ± 4%  -33.97%  (p=0.008 n=5+5)
_Pipeline/label_extractor_string-16    10.8µs ± 1%     7.0µs ± 0%  -34.90%  (p=0.016 n=5+4)

name                                 old alloc/op   new alloc/op   delta
_Pipeline/pipeline_bytes-16            9.46kB ± 0%    3.26kB ± 0%  -65.58%  (p=0.008 n=5+5)
_Pipeline/pipeline_string-16           9.46kB ± 0%    3.26kB ± 0%  -65.57%  (p=0.008 n=5+5)
_Pipeline/line_extractor_bytes-16      9.46kB ± 0%    3.26kB ± 0%  -65.58%  (p=0.008 n=5+5)
_Pipeline/line_extractor_string-16     9.46kB ± 0%    3.26kB ± 0%  -65.58%  (p=0.008 n=5+5)
_Pipeline/label_extractor_bytes-16     9.46kB ± 0%    3.26kB ± 0%  -65.58%  (p=0.008 n=5+5)
_Pipeline/label_extractor_string-16    9.46kB ± 0%    3.26kB ± 0%  -65.58%  (p=0.008 n=5+5)

name                                 old allocs/op  new allocs/op  delta
_Pipeline/pipeline_bytes-16              44.0 ± 0%      33.0 ± 0%  -25.00%  (p=0.008 n=5+5)
_Pipeline/pipeline_string-16             44.0 ± 0%      33.0 ± 0%  -25.00%  (p=0.008 n=5+5)
_Pipeline/line_extractor_bytes-16        44.0 ± 0%      33.0 ± 0%  -25.00%  (p=0.008 n=5+5)
_Pipeline/line_extractor_string-16       44.0 ± 0%      33.0 ± 0%  -25.00%  (p=0.008 n=5+5)
_Pipeline/label_extractor_bytes-16       44.0 ± 0%      33.0 ± 0%  -25.00%  (p=0.008 n=5+5)
_Pipeline/label_extractor_string-16      44.0 ± 0%      33.0 ± 0%  -25.00%  (p=0.008 n=5+5)
```